Extract fixture-loading code
